Added guards and extra functionality

+++ ToDoApplication/ToDoApp/Application.cs
@@ -42,6 +42,9 @@ namespace ToDoApp
             Console.WriteLine("Type 'e' to edit an existing item:");
             Console.WriteLine("Type 'r' to remove an item:");
             Console.WriteLine("Type 'm' to mark an item completed:");
+            Console.WriteLine("Type 'd' to display ToDo item.");
+            Console.WriteLine("Type 'p' to display ToDo List.");
+
             input = Console.ReadLine();
 
 
@@ -50,15 +53,25 @@ namespace ToDoApp
 
                 case "a":
                     AddToDoItem();
+                    PrintToDoItems();
                     break;
                 case "e":
                     EditToDoItem();
+                    PrintToDoItems();
                     break;
                 case "r":
                     RemoveToDoItem();
+                    PrintToDoItems();
                     break;
                 case "m":
                     MarkItemComplete();
+                    PrintToDoItems();
+                    break;
+                case "d":
+                    DisplayItemByID();
+                    break;
+                case "p":
+                    PrintToDoItems();
                     break;
                 default:
                     Console.WriteLine("Invalid selection.");
@@ -68,7 +81,7 @@ namespace ToDoApp
 
             }
             Console.WriteLine();
-            PrintToDoItems();
+
             Console.WriteLine("Press enter to continue.");
             Console.ReadLine();
 
@@ -76,15 +89,17 @@ namespace ToDoApp
         }
         public void PrintToDoItems()
         {
+            Console.WriteLine("Here is your current ToDo List:");
 
             foreach (var t in context.ToDOS)
             {
-
-                Console.WriteLine("Here is your current ToDo List:");
+                Console.WriteLine();
                 Console.Write("ID: {0} ", t.ID);
                 Console.WriteLine("Title: {0}  ", t.Title);
                 Console.WriteLine("ToDo: {0}", t.Content);
                 Console.WriteLine("Is Completed : {0}", t.IsComplete);
+                Console.WriteLine();
+                Console.WriteLine("~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~");
 
 
             }
@@ -102,8 +117,23 @@ namespace ToDoApp
 
             Console.WriteLine("Title: ");
             myTitle = Console.ReadLine();
+            if (myTitle == "")
+            {
+                myTitle = "Default";
+
+            }
+
+
             Console.WriteLine("Item To Do:  ");
             myContent = Console.ReadLine();
+            if (myContent == "")
+            {
+
+                myContent = "Default";
+
+            }
+
+
             myTD = new ToDo
             {
                 Title = myTitle,
@@ -134,22 +164,53 @@ namespace ToDoApp
             int.TryParse(Console.ReadLine(), out myId);
             myTD = findById(myId);
 
+            if (myTD.ID > 0)
+            {
+                Console.WriteLine("What is the new title for your item?");
+                newTitle = Console.ReadLine();
+                Console.WriteLine("What is the new thing you need to do?");
+                newContent = Console.ReadLine();
 
-            Console.WriteLine("What is the new title for your item?");
-            newTitle = Console.ReadLine();
-            Console.WriteLine("What is the new thing you need to do?");
-            newContent = Console.ReadLine();
+                myTD.Title = newTitle;
+                myTD.Content = newContent;
+                this.context.ToDOS.Update(myTD);
+                this.context.SaveChanges();
 
-            myTD.Title = newTitle;
-            myTD.Content = newContent;
-            this.context.ToDOS.Update(myTD);
-            this.context.SaveChanges();
+
+            }
+            else
+            {
+                Console.WriteLine("!!!!!!!!!!!!!!!!!");
+                Console.WriteLine("Invalid ID");
+                Console.WriteLine("!!!!!!!!!!!!!!!!!");
+            }
+
 
 
 
 
         }
 
+        public void DisplayItemByID()
+        {
+            ToDo myTD;
+            int myId;
+
+            Console.WriteLine("What item would you like to display?");
+            int.TryParse(Console.ReadLine(), out myId);
+
+            myTD = findById(myId);
+
+            Console.WriteLine("~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~");
+            Console.Write("ID:{0} ", myTD.ID);
+            Console.WriteLine("Title: {0}", myTD.Title);
+            Console.WriteLine("ToDo Item: {0}", myTD.Content);
+            Console.WriteLine("~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~");
+
+
+
+
+        }
 
         public void MarkItemComplete()
         {
@@ -159,9 +220,13 @@ namespace ToDoApp
             int.TryParse(Console.ReadLine(), out id);
 
             ToDo myTD = findById(id);
-            myTD.IsComplete = true;
-            context.ToDOS.Update(myTD);
-            context.SaveChanges();
+            if (myTD.ID > 0)
+            {
+                myTD.IsComplete = true;
+                context.ToDOS.Update(myTD);
+                context.SaveChanges();
+            }
+
 
 
         }
@@ -197,11 +262,16 @@ namespace ToDoApp
         public void RemoveToDoItem()
         {
             int myId;
+            bool itemRemoved = false;
             PrintToDoItems();
             Console.WriteLine("What is the id of the item you wish to remove?");
             int.TryParse(Console.ReadLine(), out myId);
-            removeToDo(myId);
-            Console.WriteLine("Your item has been removed.");
+            itemRemoved = removeToDo(myId);
+            if (itemRemoved)
+            {
+                Console.WriteLine("Your item has been removed.");
+            }
+
 
 
         }
@@ -214,19 +284,54 @@ namespace ToDoApp
 
         }
 
-        private void removeToDo(int id)
+        private bool removeToDo(int id)
         {
-            this.context.ToDOS.Remove(findById(id));
-            context.SaveChanges();
+            bool removed = false;
+            if (findById(id).ID != -1)
+            {
+                try
+                {
+                    this.context.ToDOS.Remove(findById(id));
+                    context.SaveChanges();
+                    removed = true;
+
+                }
+                catch (System.InvalidOperationException ex)
+                {
+                    Console.WriteLine("!!!!!!!!!!!!!!!!!");
+                    throw new System.Exception("!!!!!ID not found.!!!!!", ex);
+
+                }
+
+            }
+            else
+            {
+                Console.WriteLine("!!!!!!!!!!!!!!!!!");
+                Console.WriteLine("Not a valid value");
+                Console.WriteLine("!!!!!!!!!!!!!!!!!");
+            }
+
+            return removed;
 
         }
+
+
 
         private ToDo findById(int id)
         {
             ToDo td;
             td = this.context.ToDOS.Find(id);
-            // var entity = this.context.ToDOS.FirstOrDefaultAsync(item => item.ID == id);
-            return td;
+            if (td != null)
+            { return td; }
+
+            else
+            {
+                td = new ToDo { ID = -1, Title = "", Content = "", CreateDate = DateTime.Now, IsComplete = false };
+                Console.WriteLine("!!!!!!!!!!!!!!!!!");
+                Console.WriteLine("Item not found.");
+                Console.WriteLine("!!!!!!!!!!!!!!!!!");
+                return td;
+            }
 
         }
     }
